requirements.txt
fastapi
uvicorn
pyyaml
pandas
numpy
matplotlib
